* [html5] add funciontal methods createEventMap & trimTextNodes.
