+++ html5/render/vue/components/scrollable/header.js
@@ -1,6 +1,6 @@
 // import { validateStyles } from '../../validator'
 import { supportSticky } from '../../utils/style'
-import { extractComponentStyle } from '../../core'
+import { extractComponentStyle, createEventMap } from '../../core'
 
 export default {
   data () {
@@ -49,7 +49,7 @@ export default {
     // }
     return createElement('html:header', {
       attrs: { 'weex-type': 'header' },
-      on: this._createEventMap(),
+      on: createEventMap(this),
       ref: 'header',
       staticClass: 'weex-header weex-ct',
       class: { sticky: this.sticky, iossticky: this.supportSticky },
