+++ html5/render/vue/components/video.js
@@ -1,5 +1,5 @@
 // import { validateStyles } from '../validator'
-import { extractComponentStyle } from '../core'
+import { extractComponentStyle, createEventMap } from '../core'
 
 export default {
   props: {
@@ -46,7 +46,7 @@ export default {
         controls: this.controls,
         src: this.src
       },
-      on: this._createEventMap(['start', 'pause', 'finish', 'fail']),
+      on: createEventMap(this, ['start', 'pause', 'finish', 'fail']),
       staticClass: 'weex-video weex-el',
       staticStyle: extractComponentStyle(this)
     })
