+++ html5/render/vue/core/index.js
@@ -1,1 +1,2 @@
 export * from './style'
+export * from './node'
