+++ html5/test/render/vue/core/style.js
@@ -36,9 +36,8 @@ init('core style', (Vue, helper) => {
 
   it('should get normalized merged styles.', function () {
     const vm = helper.createVm(scopedStyleBundle)
-    const el = vm.$refs.foo.$el
+    const el = vm.$refs.foo.$el || vm.$refs.foo
     expect(el).to.be.ok
-
     const expectedMap = {
       width: 200 * scale + 'px',
       height: 200 * scale + 'px',
@@ -55,19 +54,19 @@ init('core style', (Vue, helper) => {
       transform: `translate3d(${100 * scale + 'px'}, ${100 * scale + 'px'}, 0px)`
     }
 
-   for (const k in expectedMap) {
-    expect(el.style[k]).to.equal(expectedMap[k])
-   }
-   const directionRes = []
-   for (const k in expectedDirection) {
-    directionRes.push(el.style[k] === expectedDirection[k])
-   }
-   expect(directionRes).to.include(true)
+    for (const k in expectedMap) {
+      expect(el.style[k]).to.equal(expectedMap[k])
+    }
+    const directionRes = []
+    for (const k in expectedDirection) {
+      directionRes.push(el.style[k] === expectedDirection[k])
+    }
+    expect(directionRes).to.include(true)
 
-   const transformRes = []
-   for (const k in expectedTransform) {
-    transformRes.push(el.style[k] === expectedTransform[k])
-   }
-   expect(transformRes).to.include(true)
+    const transformRes = []
+    for (const k in expectedTransform) {
+      transformRes.push(el.style[k] === expectedTransform[k])
+    }
+    expect(transformRes).to.include(true)
   })
 })
